Move the migration in the right polyfill
