--- src/Polyfill48/DependencyInjection/CcaContaoPolyfill48Extension.php
@@ -1,51 +1,0 @@
-<?php
-
-/**
- * This file is part of contao-community-alliance/contao-polyfill-bundle.
- *
- * (c) 2019-2020 Contao Community Alliance.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    contao-community-alliance/contao-polyfill-bundle
- * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2019-2020 Contao Community Alliance.
- * @license    https://github.com/contao-community-alliance/contao-polyfill-bundle/blob/master/LICENSE LGPL-3.0-or-later
- * @filesource
- */
-
-declare(strict_types = 1);
-
-namespace ContaoCommunityAlliance\Polyfills\Polyfill48\DependencyInjection;
-
-use ContaoCommunityAlliance\Polyfills\Polyfill48\Migration\MigrationInterfacePolyFill;
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Extension\Extension;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-
-/**
- * The extension for load configuration.
- */
-final class CcaContaoPolyfill48Extension extends Extension
-{
-    /**
-     * {@inheritDoc}
-     */
-    public function load(array $configs, ContainerBuilder $container): void
-    {
-        $config = $this->processConfiguration($this->getConfiguration($configs, $container), $configs);
-        $loader = new YamlFileLoader($container, new FileLocator(\dirname(__DIR__) . '/Resources/config'));
-
-        if ($config['migration']) {
-            $loader->load('migration.yml');
-
-            $container
-                ->registerForAutoconfiguration(MigrationInterfacePolyFill::class)
-                ->addTag('contao.migration');
-        }
-    }
-}
