--- src/Polyfill48/DependencyInjection/Configuration.php
@@ -1,55 +1,0 @@
-<?php
-
-/**
- * This file is part of contao-community-alliance/contao-polyfill-bundle.
- *
- * (c) 2019-2020 Contao Community Alliance.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    contao-community-alliance/contao-polyfill-bundle
- * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2019-2020 Contao Community Alliance.
- * @license    https://github.com/contao-community-alliance/contao-polyfill-bundle/blob/master/LICENSE LGPL-3.0-or-later
- * @filesource
- */
-
-declare(strict_types = 1);
-
-namespace ContaoCommunityAlliance\Polyfills\Polyfill48\DependencyInjection;
-
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\Config\Definition\Builder\TreeBuilder;
-use Symfony\Component\Config\Definition\ConfigurationInterface;
-
-/**
- * The configuration.
- */
-final class Configuration implements ConfigurationInterface
-{
-    /**
-     * {@inheritDoc}
-     */
-    public function getConfigTreeBuilder(): TreeBuilder
-    {
-        $treeBuilder = new TreeBuilder('cca_polyfill48');
-        /** @var ArrayNodeDefinition $rootNode */
-        $rootNode = \method_exists($treeBuilder, 'getRootNode')
-            ? $treeBuilder->getRootNode()
-                : $treeBuilder->root('cca_polyfill48');
-
-        $rootNode
-            ->addDefaultsIfNotSet()
-            ->children()
-                ->booleanNode('migration')
-                    ->info('Use the Contao migration service.')
-                    ->defaultTrue()
-                ->end()
-            ->end();
-
-        return $treeBuilder;
-    }
-}
