+++ src/Polyfill49/Resources/ide-config.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * This file is part of contao-community-alliance/contao-polyfill-bundle.
+ *
+ * (c) 2019-2020 Contao Community Alliance.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * This project is provided in good faith and hope to be usable by anyone.
+ *
+ * @package    contao-community-alliance/contao-polyfill-bundle
+ * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @copyright  2019-2020 Contao Community Alliance.
+ * @license    https://github.com/contao-community-alliance/contao-polyfill-bundle/blob/master/LICENSE LGPL-3.0-or-later
+ * @filesource
+ */
+
+// This file is only here to allow IDE auto-completion.
+// @codingStandardsIgnoreFile
+
+namespace Contao\CoreBundle\Migration {
+    if (false) {
+        interface MigrationInterface extends \ContaoCommunityAlliance\Polyfills\Polyfill49\Migration\MigrationInterfacePolyFill {}
+        class MigrationResult extends \ContaoCommunityAlliance\Polyfills\Polyfill49\Migration\MigrationResultPolyFill {}
+        class MigrationCollection extends \ContaoCommunityAlliance\Polyfills\Polyfill49\Migration\MigrationCollectionPolyFill {}
+    }
+}
