+++ tests/Polyfill49/Fixtures/Updater.php
@@ -1,0 +1,43 @@
+<?php
+
+/**
+ * This file is part of contao-community-alliance/contao-polyfill-bundle.
+ *
+ * (c) 2019-2020 Contao Community Alliance.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * This project is provided in good faith and hope to be usable by anyone.
+ *
+ * @package    contao-community-alliance/contao-polyfill-bundle
+ * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @copyright  2019-2020 Contao Community Alliance.
+ * @license    https://github.com/contao-community-alliance/contao-polyfill-bundle/blob/master/LICENSE LGPL-3.0-or-later
+ * @filesource
+ */
+
+declare(strict_types=1);
+
+namespace ContaoCommunityAlliance\Polyfills\Test\Polyfill49\Fixtures;
+
+use Contao\InstallationBundle\Database\AbstractVersionUpdate;
+
+class Updater extends AbstractVersionUpdate
+{
+    /**
+     * @inheritDoc
+     */
+    public function shouldBeRun()
+    {
+        // TODO: Implement shouldBeRun() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function run()
+    {
+        // TODO: Implement run() method.
+    }
+}
